backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.42
psycopg[binary]==3.2.9
alembic==1.16.5
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
apscheduler==3.11.0
httpx==0.28.1
python-dateutil==2.9.0.post0
rapidfuzz==3.14.0
cryptography==45.0.6
openai==1.101.0
google-generativeai==0.8.5
python-docx==1.2.0
pypdf==6.0.0
reportlab==4.4.3
Pillow==11.3.0
redis==6.4.0
